simplicity ftw
btw in the course of hacking this, I discovered that Hacker News
doesn't just use tables within tables, it also uses invisible spacer
gifs. this is a technique so old I think there's a safe bet most
Hacker News readers have never even heard of it. it was state of the
art in 1996.