stack overflow's stubborn overlay header
I think this may be more useful in terms of removing it, but it's
actually impossible to say for sure, as I don't know the rules
behind when it appears in the first place. it's a "welcome back!"
thing which is based on some very obviously inaccurate measure of
how often Stack Overflow imagines me to be visiting the site.

anyway, this might work. if not I may have to investigate more
aggressive measures.