removes slide-over sign up bar
give it up, quora, it's never going to happen. I'm not ever going
to sign up or log in. you're a site which tries to be a Stack
Overflow for everything but code, and where a question about side
projects rates a comment BY A GUY WHO BUILT A FUCKING PLANE below
a comment from one of your own engineers who supplies no evidence
of having ever completed any side project successfully at all.

http://www.quora.com/Productivity/How-can-I-increase-my-productivity-on-my-side-projects-at-the-end-of-the-day-when-Im-tired-from-work

reminds me of this:

http://www.youtube.com/watch?v=bExOGt34ZIU