ironically this blog advocates for better CSS
this is really funny. a Microsoft platforms guy writes a blog post about
writing better JS and CSS. I read it because it's on HN and read his
CSS/DOM shit so that I can clear out the sidebar(s) and make his shit
readable. the better CSS he advocates writing does not appear in the
code for his blog. just sayin.